Sync latest from grafana/main to main

+++ pkg/services/contexthandler/contexthandler.go
@@ -22,8 +22,11 @@ import (
 	"github.com/grafana/grafana/pkg/middleware/cookies"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/apikey"
+	"github.com/grafana/grafana/pkg/services/auth"
+	"github.com/grafana/grafana/pkg/services/authn"
 	"github.com/grafana/grafana/pkg/services/contexthandler/authproxy"
 	"github.com/grafana/grafana/pkg/services/contexthandler/ctxkey"
+	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/services/login"
 	"github.com/grafana/grafana/pkg/services/oauthtoken"
 	"github.com/grafana/grafana/pkg/services/org"
@@ -42,11 +45,12 @@ const (
 
 const ServiceName = "ContextHandler"
 
-func ProvideService(cfg *setting.Cfg, tokenService models.UserTokenService, jwtService models.JWTService,
+func ProvideService(cfg *setting.Cfg, tokenService auth.UserTokenService, jwtService models.JWTService,
 	remoteCache *remotecache.RemoteCache, renderService rendering.Service, sqlStore db.DB,
 	tracer tracing.Tracer, authProxy *authproxy.AuthProxy, loginService login.Service,
 	apiKeyService apikey.Service, authenticator loginpkg.Authenticator, userService user.Service,
-	orgService org.Service, oauthTokenService oauthtoken.OAuthTokenService,
+	orgService org.Service, oauthTokenService oauthtoken.OAuthTokenService, features *featuremgmt.FeatureManager,
+	authnService authn.Service,
 ) *ContextHandler {
 	return &ContextHandler{
 		Cfg:               cfg,
@@ -63,13 +67,14 @@ func ProvideService(cfg *setting.Cfg, tokenService models.UserTokenService, jwtS
 		userService:       userService,
 		orgService:        orgService,
 		oauthTokenService: oauthTokenService,
+		features:          features,
 	}
 }
 
 // ContextHandler is a middleware.
 type ContextHandler struct {
 	Cfg               *setting.Cfg
-	AuthTokenService  models.UserTokenService
+	AuthTokenService  auth.UserTokenService
 	JWTAuthService    models.JWTService
 	RemoteCache       *remotecache.RemoteCache
 	RenderService     rendering.Service
@@ -82,6 +87,8 @@ type ContextHandler struct {
 	userService       user.Service
 	orgService        org.Service
 	oauthTokenService oauthtoken.OAuthTokenService
+	features          *featuremgmt.FeatureManager
+	authnService      authn.Service
 	// GetTime returns the current time.
 	// Stubbable by tests.
 	GetTime func() time.Time
@@ -179,17 +186,33 @@ func (h *ContextHandler) Middleware(next http.Handler) http.Handler {
 			}
 		}
 
+		// this can be used by proxies to identify certain users
+		if h.features.IsEnabled(featuremgmt.FlagReturnUnameHeader) {
+			w.Header().Add("grafana-uname", reqContext.Login)
+		}
+
 		next.ServeHTTP(w, r)
 	})
 }
 
 func (h *ContextHandler) initContextWithAnonymousUser(reqContext *models.ReqContext) bool {
+	ctx, span := h.tracer.Start(reqContext.Req.Context(), "initContextWithAnonymousUser")
+	defer span.End()
+
+	if h.features.IsEnabled(featuremgmt.FlagAuthnService) {
+		identity, err := h.authnService.Authenticate(ctx, authn.ClientAnonymous, &authn.Request{HTTPRequest: reqContext.Req})
+		if err != nil {
+			return false
+		}
+		reqContext.SignedInUser = identity.SignedInUser()
+		reqContext.IsSignedIn = false
+		reqContext.AllowAnonymous = true
+		return true
+	}
+
 	if !h.Cfg.AnonymousEnabled {
 		return false
 	}
-
-	_, span := h.tracer.Start(reqContext.Req.Context(), "initContextWithAnonymousUser")
-	defer span.End()
 
 	getOrg := org.GetOrgByNameQuery{Name: h.Cfg.AnonymousOrgName}
 
@@ -425,9 +448,12 @@ func (h *ContextHandler) initContextWithToken(reqContext *models.ReqContext, org
 
 	token, err := h.AuthTokenService.LookupToken(ctx, rawToken)
 	if err != nil {
-		reqContext.Logger.Warn("Failed to look up user based on cookie", "error", err)
-		// Burn the cookie in case of failure
-		reqContext.Resp.Before(h.deleteInvalidCookieEndOfRequestFunc(reqContext))
+		reqContext.Logger.Warn("failed to look up session from cookie", "error", err)
+		if errors.Is(err, auth.ErrUserTokenNotFound) || errors.Is(err, auth.ErrInvalidSessionToken) {
+			// Burn the cookie in case of invalid, expired or missing token
+			reqContext.Resp.Before(h.deleteInvalidCookieEndOfRequestFunc(reqContext))
+		}
+
 		reqContext.LookupTokenErr = err
 
 		return false
@@ -440,34 +466,30 @@ func (h *ContextHandler) initContextWithToken(reqContext *models.ReqContext, org
 		return false
 	}
 
-	getTime := h.GetTime
-	if getTime == nil {
-		getTime = time.Now
-	}
+	if h.features.IsEnabled(featuremgmt.FlagAccessTokenExpirationCheck) {
+		// Check whether the logged in User has a token (whether the User used an OAuth provider to login)
+		oauthToken, exists, _ := h.oauthTokenService.HasOAuthEntry(ctx, queryResult)
+		if exists {
+			if h.hasAccessTokenExpired(oauthToken) {
+				reqContext.Logger.Info("access token expired", "userId", query.UserID, "expiry", fmt.Sprintf("%v", oauthToken.OAuthExpiry))
 
-	// Check whether the logged in User has a token (whether the User used an OAuth provider to login)
-	oauthToken, exists, _ := h.oauthTokenService.HasOAuthEntry(ctx, queryResult)
-	if exists {
-		// Skip where the OAuthExpiry is default/zero/unset
-		if !oauthToken.OAuthExpiry.IsZero() && oauthToken.OAuthExpiry.Round(0).Add(-oauthtoken.ExpiryDelta).Before(getTime()) {
-			reqContext.Logger.Info("access token expired", "userId", query.UserID, "expiry", fmt.Sprintf("%v", oauthToken.OAuthExpiry))
+				// If the User doesn't have a refresh_token or refreshing the token was unsuccessful then log out the User and invalidate the OAuth tokens
+				if err = h.oauthTokenService.TryTokenRefresh(ctx, oauthToken); err != nil {
+					if !errors.Is(err, oauthtoken.ErrNoRefreshTokenFound) {
+						reqContext.Logger.Error("could not fetch a new access token", "userId", oauthToken.UserId, "error", err)
+					}
 
-			// If the User doesn't have a refresh_token or refreshing the token was unsuccessful then log out the User and Invalidate the OAuth tokens
-			if err = h.oauthTokenService.TryTokenRefresh(ctx, oauthToken); err != nil {
-				if !errors.Is(err, oauthtoken.ErrNoRefreshTokenFound) {
-					reqContext.Logger.Error("could not fetch a new access token", "userId", oauthToken.UserId, "error", err)
+					reqContext.Resp.Before(h.deleteInvalidCookieEndOfRequestFunc(reqContext))
+					if err = h.oauthTokenService.InvalidateOAuthTokens(ctx, oauthToken); err != nil {
+						reqContext.Logger.Error("could not invalidate OAuth tokens", "userId", oauthToken.UserId, "error", err)
+					}
+
+					err = h.AuthTokenService.RevokeToken(ctx, token, false)
+					if err != nil && !errors.Is(err, auth.ErrUserTokenNotFound) {
+						reqContext.Logger.Error("failed to revoke auth token", "error", err)
+					}
+					return false
 				}
-
-				reqContext.Resp.Before(h.deleteInvalidCookieEndOfRequestFunc(reqContext))
-				if err = h.oauthTokenService.InvalidateOAuthTokens(ctx, oauthToken); err != nil {
-					reqContext.Logger.Error("could not invalidate OAuth tokens", "userId", oauthToken.UserId, "error", err)
-				}
-
-				err = h.AuthTokenService.RevokeToken(ctx, token, false)
-				if err != nil && !errors.Is(err, models.ErrUserTokenNotFound) {
-					reqContext.Logger.Error("failed to revoke auth token", "error", err)
-				}
-				return false
 			}
 		}
 	}
@@ -495,8 +517,8 @@ func (h *ContextHandler) deleteInvalidCookieEndOfRequestFunc(reqContext *models.
 	}
 }
 
-func (h *ContextHandler) rotateEndOfRequestFunc(reqContext *models.ReqContext, authTokenService models.UserTokenService,
-	token *models.UserToken) web.BeforeFunc {
+func (h *ContextHandler) rotateEndOfRequestFunc(reqContext *models.ReqContext, authTokenService auth.UserTokenService,
+	token *auth.UserToken) web.BeforeFunc {
 	return func(w web.ResponseWriter) {
 		// if response has already been written, skip.
 		if w.Written() {
@@ -720,4 +742,17 @@ func AuthHTTPHeaderListFromContext(c context.Context) *AuthHTTPHeaderList {
 		return list
 	}
 	return nil
+}
+
+func (h *ContextHandler) hasAccessTokenExpired(token *models.UserAuth) bool {
+	if token.OAuthExpiry.IsZero() {
+		return false
+	}
+
+	getTime := h.GetTime
+	if getTime == nil {
+		getTime = time.Now
+	}
+
+	return token.OAuthExpiry.Round(0).Add(-oauthtoken.ExpiryDelta).Before(getTime())
 }
